// sanity/types.ts
export type Project = {
  _id: string;
  _type: "project";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  position?: number;
  date?: string;
  name?: string;
  description?: string;
  type?: string;
  location?: string;
  link?: string;
  slug?: Slug;
  views?: number;
};

export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};